Don't unescape when splitting.
The unescaping happens when we serialize the Segment. It shouldn't
happen here.

This fixes the following issue:
A value that ends in a '?' is properly escaped to '??'. When parsing the
value, we split and take away one of the '?' so we're just left with
'?'. This now becomes an invalid value because we have a dangling escape
character.